+++ rust/gst-nmos-rs/src/session/mod.rs
@@ -18,10 +18,11 @@ use gstreamer as gst;
 use nvnmos_rpc::v1::Transport as ProtoTransport;
 
 use crate::daemon::{ActivationHandler, ActivationRequest, Session};
-use crate::domain::{self, DomainIdOrigin};
-use crate::flow_def::{self, FlowDefBuildInput, FlowDefOverrides, ValueOrigin};
+use crate::domain;
+use crate::flow_def::{self, FlowDefBuildInput};
 use crate::runtime::SHARED_RUNTIME;
-use crate::sdp::{self, SdpOverrides};
+#[cfg(test)]
+use crate::sdp;
 use crate::types::{CapsMode, FlowFormat, Transport};
 
 /// Open-session timeout. Aligned with the daemon's activation ack
@@ -481,115 +482,19 @@ impl TransportConfig {
     }
 }
 
-/// Which factory family to use for the UDP socket and RTP
-/// (de)payloader elements.
-///
-/// `V1` is gst-plugins-good throughout: `udpsrc` / `udpsink` /
-/// `rtpvrawpay` / `rtpL24pay` / etc. `V2` prefers gst-plugins-rs
-/// (`udpsrc2`, `rtpL24pay2`, `rtpL24depay2`, …) on a per-element
-/// basis and falls back to the V1 factory for any element that
-/// doesn't yet have a V2 sibling. Per-element fallback (rather
-/// than all-or-nothing) matters because the V2 family is rolled
-/// out incrementally upstream — for example today
-/// gst-plugins-rs ships `rtpL24pay2`/`depay2` but not yet
-/// `rtpvrawpay2`/`depay2`, and no `udpsink2` exists at all (the
-/// performance motivation for `udpsrc2` was kernel receive
-/// efficiency, which doesn't translate to the send side).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum UdpVariant {
-    V1,
-    V2,
-}
+mod mxl;
+mod udp;
 
-/// One RTP media line's worth of state — everything the UDP chain
-/// factories need to instantiate the inner elements.
-///
-/// Essence-level state (`format`, `rtp_caps`, `raw_caps`) is shared
-/// across legs because both legs of an ST 2022-7 pair carry the
-/// same essence with the same PT / clock-rate / encoding-name; only
-/// the network params differ. Per-leg state lives on [`UdpLeg`].
-///
-/// Field names use NMOS / IS-05 terminology (`destination_ip`,
-/// `interface_ip`, `source_ip`, ...) for direction independence;
-/// the public element properties on `nmossrc` / `nmossink` use the
-/// IS-05 RTP transport_params vocabulary verbatim (`source-ip`,
-/// `source-port`, `destination-ip`, `destination-port`,
-/// `interface-ip`, `multicast-ip`), mapped onto these per-leg
-/// fields at property-set / SDP-splice time — see
-/// [`CommonSettings::source_ip`] et seq. for the per-side wire
-/// semantics. The mapping is 1:1 to IS-05 wire JSON, so a
-/// controller PATCHing `/single/senders/{id}/staged` reads
-/// straight into the same GObject property names. How the
-/// redundant secondary leg gets exposed on the property surface
-/// is a separate design decision — `nvdsudpsrc` for example
-/// overloads `local-iface-ip` into a comma-separated list and
-/// adds a combined `st2022-7-streams` property rather than
-/// `-2`-suffixed scalar twins — and is deferred until the
-/// redundancy work lands.
-#[derive(Debug, Clone)]
-pub(crate) struct UdpMedia {
-    /// Essence family — selects the payloader / depayloader factory
-    /// alongside [`UdpVariant`].
-    pub(crate) format: FlowFormat,
-    /// First (and, for non-redundant RTP, only) leg.
-    pub(crate) primary: UdpLeg,
-    /// Redundant secondary leg for ST 2022-7. `None` for
-    /// non-redundant RTP — which is everything today, until the
-    /// 2022-7 work lands.
-    pub(crate) secondary: Option<UdpLeg>,
-    /// `application/x-rtp,...` caps the depayloader consumes (and
-    /// the payloader produces). Carries PT, clock-rate,
-    /// encoding-name, channels, sampling, depth and any other
-    /// essence-specific RFC 4175 / RFC 3551 / RFC 3190 parameters
-    /// that `a=rtpmap` / `a=fmtp` map to. `a=ptime:` / `a=maxptime:`
-    /// are hoisted onto these caps as `a-ptime` / `a-maxptime`
-    /// (the GStreamer convention `SDPMedia::set_media_from_caps`
-    /// rebuilds into standalone `a=…:` SDP attributes). The
-    /// payloader / depayloader and the chain factories
-    /// ([`crate::inner::build_udpsink`] et al) read this field
-    /// directly.
-    pub(crate) rtp_caps: gst::Caps,
-    /// Essence caps (`video/x-raw,…`, `audio/x-raw,…`,
-    /// `meta/x-st-2038,…`). The receiver pins these on its ghost
-    /// src pad so downstream caps queries see the concrete shape
-    /// the flow will carry, mirroring the MXL path's
-    /// `advertise_caps` derived from the flow_def.
-    pub(crate) raw_caps: gst::Caps,
-}
+pub(crate) use udp::{UdpLeg, UdpMedia, UdpVariant};
+pub(crate) use mxl::decide_inner_config_mxl;
+#[cfg(test)]
+pub(crate) use mxl::{property_overrides_mxl, synthesise_or_passthrough_mxl};
+pub(crate) use udp::{decide_inner_config_udp, udp_essence_cross_check_mode};
+#[cfg(test)]
+pub(crate) use udp::property_overrides_udp;
 
-/// One network leg of a [`UdpMedia`]. Non-redundant RTP has a single
-/// leg ([`UdpMedia::primary`]); ST 2022-7 adds a second
-/// ([`UdpMedia::secondary`]) carrying the same essence over an
-/// independent network path for hitless merging.
-///
-/// All fields are per-leg state that NMOS IS-05's
-/// `transport_params` carries one-for-one (with `source_ip`
-/// modelled as the NMOS-simplified single-entry equivalent of the
-/// SDP `a=source-filter:` include list — see field doc).
-#[derive(Debug, Clone)]
-pub(crate) struct UdpLeg {
-    /// Multicast group (or unicast destination). Sender's
-    /// `udpsink.host` / receiver's `udpsrc.address`.
-    pub(crate) destination_ip: String,
-    /// Sender's `udpsink.port` / receiver's `udpsrc.port`.
-    pub(crate) destination_port: u16,
-    /// Local interface IP. Nvds elements take this directly as
-    /// `local-iface-ip`; for `udpsrc` / `udpsink` we resolve to an
-    /// interface name and forward as `multicast-iface`.
-    pub(crate) interface_ip: Option<String>,
-    /// SSM source-IP filter. Receiver-only. NMOS-RTP
-    /// `transport_params[i].source_ip` is a single string by
-    /// design — the SDP `a=source-filter:` line supports list /
-    /// exclude semantics but NMOS constrains itself to one
-    /// include-mode source per leg. We forward this directly to
-    /// `nvdsudpsrc.source-address`; on the gst-plugins-good
-    /// `udpsrc` path it's advertised in NMOS but not currently
-    /// enforced at the socket (no native source-filter property).
-    pub(crate) source_ip: Option<String>,
-    /// Sender source port. Forwarded as `udpsink.bind-port`.
-    /// Sender-only.
-    pub(crate) source_port: Option<u16>,
-}
+use mxl::{resolve_activation_inner_mxl, resolve_inner_config_mxl};
+use udp::resolve_inner_config_udp;
 
 /// Validate the settings snapshot and open a session via the shared
 /// tokio runtime. On success the session is stored under `session`
@@ -723,593 +628,13 @@ pub(crate) fn validate_and_open(
     *session.lock().unwrap() = Some(new_session);
     Ok(inner)
 }
-
-/// If the user supplied a `transport-file` (literal or path), pass
-/// it through; otherwise, when `caps` is set and `mxl-flow-id` is
-/// non-empty, synthesise a flow_def JSON document via
-/// [`flow_def::from_caps`]. When *both* a transport file and
-/// `caps` are set, the file is passed through and the caps are
-/// cross-checked against the file's `format` further down the
-/// validate path (a mismatch is a hard error, not silently dropped).
-///
-/// Senders and Receivers both go through synthesis: a Sender's
-/// `flow_def` describes the Flow it produces; a Receiver's *configuring*
-/// `flow_def` describes the essence shape this Receiver is configured
-/// to accept (BCP-004-01 narrow Receiver Caps), with the `urn:x-nvnmos:tag:caps`
-/// tag spliced in by `receiver-caps-mode` to indicate narrow vs wide.
-/// The live transport file that arrives via IS-05 PATCH replaces the
-/// subscription-relevant fields (mxl-flow-id, etc.) at activation time
-/// but the configuring file is what the daemon uses for IS-04
-/// advertisement at registration time.
-///
-/// Returns `Ok(None)` when nothing can be synthesised — neither a
-/// transport file nor enough property state to build one. The element
-/// then opens the session without a transport file and runs on the
-/// fake chain until an IS-05 activation arrives.
-fn synthesise_or_passthrough_mxl(
-    cat: &gst::DebugCategory,
-    element: &str,
-    settings: &CommonSettings,
-    resolved_mxl_domain_id: &str,
-    resolved: Option<String>,
-) -> Result<Option<String>, anyhow::Error> {
-    match (resolved, settings.caps.as_ref()) {
-        (Some(text), Some(_)) => {
-            gst::debug!(
-                cat,
-                "{element}: transport-file set; `caps` will be cross-checked against the file's `format`"
-            );
-            Ok(Some(text))
-        }
-        (Some(text), None) => Ok(Some(text)),
-        (None, Some(caps)) => {
-            if settings.mxl_flow_id.is_empty() {
-                gst::debug!(
-                    cat,
-                    "{element}: `caps` set but `mxl-flow-id` empty; deferring flow_def \
-                     synthesis (the fake chain will be in use until an IS-05 \
-                     activation supplies the flow id)"
-                );
-                return Ok(None);
-            }
-            let json = flow_def::from_caps(&FlowDefBuildInput {
-                flow_id: &settings.mxl_flow_id,
-                name: &settings.name,
-                mxl_domain_id: resolved_mxl_domain_id,
-                label: &settings.label,
-                description: &settings.description,
-                caps,
-            })
-            .with_context(|| format!("{element}: synthesising flow_def from caps"))?;
-            gst::info!(
-                cat,
-                "{element}: synthesised flow_def from `caps` (side={:?})",
-                settings.side,
-            );
-            Ok(Some(json))
-        }
-        (None, None) => Ok(None),
-    }
-}
-
-/// UDP counterpart of [`synthesise_or_passthrough_mxl`]:
-/// produce a configuring SDP at startup time, either by
-/// passing through a user-supplied `transport-file*` or by
-/// synthesising one from `caps` + `transport-caps` +
-/// IS-05 endpoint properties + [`sdp::defaults`].
-///
-/// Returns `Ok(Some(text))` when either path produces an SDP,
-/// `Ok(None)` when neither is possible (no transport file
-/// **and** no `caps`) — the element then opens the session
-/// without one and runs on the fake chain until an IS-05
-/// activation arrives.
-///
-/// Precedence is identical to the MXL path: an explicit
-/// transport file beats `caps`-driven synthesis, mirroring
-/// the user spec "transport-file* takes priority over caps
-/// at startup; activation-SDP wins over both at PATCH time
-/// (handled in [`make_activation_plan`])".
-///
-/// Caps-only synthesis composes [`sdp::from_caps`] with the
-/// CommonSettings snapshot, mapping the per-side IS-05
-/// property vocabulary (Sender's `destination_ip` vs
-/// Receiver's `multicast_ip`; Sender's `source_ip`-as-NIC vs
-/// Receiver's distinct `interface_ip`) into the unified
-/// [`sdp::SdpBuildInput`] vocabulary that
-/// [`sdp::from_caps`] consumes.
-fn synthesise_or_passthrough_udp(
-    cat: &gst::DebugCategory,
-    element: &str,
-    settings: &CommonSettings,
-    resolved: Option<String>,
-) -> Result<Option<String>, anyhow::Error> {
-    match (resolved, settings.caps.as_ref()) {
-        (Some(text), Some(_)) => {
-            gst::debug!(
-                cat,
-                "{element}: SDP transport-file set; `caps` will be cross-checked against the file's essence shape",
-            );
-            Ok(Some(text))
-        }
-        (Some(text), None) => Ok(Some(text)),
-        (None, Some(essence_caps)) => {
-            // Per-side dispatch on the IS-05 destination slot:
-            // Senders carry it as `destination_ip`, Receivers
-            // as `multicast_ip`. The single wire slot on the
-            // SDP `m=` / `c=` line is named `destination_ip`
-            // on `sdp::SdpBuildInput`.
-            let destination_ip = match settings.side {
-                Side::Sender => settings.destination_ip.as_str(),
-                Side::Receiver => settings.multicast_ip.as_str(),
-            };
-            let advertise_caps = match settings.caps_mode {
-                // Auto resolves to narrow for synthesised
-                // SDPs — the splice path can promote to wide
-                // later if `receiver-caps-mode=Wide` is set.
-                // (For synthesis we always know
-                // `caps_mode` directly, so we can apply it
-                // here without going through the splice's
-                // text-rewrite.)
-                CapsMode::Auto | CapsMode::Narrow => false,
-                CapsMode::Wide => true,
-            };
-            let input = sdp::SdpBuildInput {
-                essence_caps,
-                transport_caps: settings.transport_caps.as_ref(),
-                side: settings.side,
-                label: &settings.label,
-                description: &settings.description,
-                name: &settings.name,
-                source_ip: &settings.source_ip,
-                source_port: settings.source_port,
-                destination_ip,
-                destination_port: settings.destination_port,
-                interface_ip: &settings.interface_ip,
-                advertise_caps,
-                node_seed: &settings.node_seed,
-            };
-            let text = sdp::from_caps(&input)
-                .with_context(|| format!("{element}: synthesising SDP from caps"))?;
-            gst::info!(
-                cat,
-                "{element}: synthesised SDP from `caps` (side={:?})",
-                settings.side,
-            );
-            Ok(Some(text))
-        }
-        (None, None) => Ok(None),
-    }
-}
-
-/// Best-effort [`FlowFormat`] derived from the `caps` property.
-/// Returns [`FlowFormat::Unspecified`] when `caps` is unset or the
-/// first structure's media type isn't one of the recognised essence
-/// shapes — the caller then falls through to the transport file's
-/// `format` (if present) or to the fake chain.
-fn caps_format(settings: &CommonSettings) -> FlowFormat {
+pub(super) fn caps_format(settings: &CommonSettings) -> FlowFormat {
     settings
         .caps
         .as_ref()
         .map(FlowFormat::from_caps)
         .unwrap_or(FlowFormat::Unspecified)
 }
-
-/// Build an [`SdpOverrides`] from the element's property snapshot
-/// for the RTP transports. Mirrors [`property_overrides_mxl`] on the
-/// MXL path:
-///
-/// * Empty-string properties map to `None` (i.e. "user did not set
-///   this; leave the file's value alone").
-/// * Zero ports map to `None` (`http-port`'s "unset" sentinel
-///   convention, extended to all u16 port slots).
-/// * `caps_mode` is passed through by value because
-///   [`CapsMode::Auto`] is already the "no override" sentinel.
-///
-/// Per-side dispatch reflects the IS-05 schema's asymmetry:
-///
-/// * **Sender**: `settings.source_ip` is the local egress NIC IP
-///   and feeds **both** [`SdpOverrides::source_ip`] (used by the
-///   splice as the SDP `a=source-filter:` include-source — RFC
-///   4607 SSM convention) and [`SdpOverrides::interface_ip`]
-///   (used as `a=x-nvnmos-iface-ip:`). The two SDP slots carry
-///   the same IP on a Sender because they're both saying "this
-///   is where the sender egresses from", just for different
-///   downstream consumers (SSM-aware receivers vs libnvnmos).
-///   Receiver-only slots (`interface_ip`, `multicast_ip`) are
-///   not read.
-/// * **Receiver**: [`SdpOverrides::source_ip`] is the SSM
-///   include-source (remote sender's IP) from
-///   `settings.source_ip`; [`SdpOverrides::interface_ip`] is the
-///   local NIC from `settings.interface_ip`;
-///   [`SdpOverrides::destination_ip`] is the multicast group
-///   from `settings.multicast_ip` (the SDP `c=` line wire slot
-///   that IS-05 splits between sender's `destination_ip` and
-///   receiver's `multicast_ip` per resource direction).
-///   `source_port` is left unset (the IS-05 receiver schema
-///   doesn't define a local source-port slot).
-///
-/// `transport_caps` populates the override-class slots
-/// (`payload_type`, `audio_clock_rate`, `a_ptime`,
-/// `a_maxptime`) by reading the corresponding fields from
-/// `application/x-rtp,...` caps. Out-of-range / missing values
-/// drop silently here — pt's RFC 3551 range check fires
-/// downstream in [`sdp::passthrough_with_overrides`], where the error
-/// can be attributed cleanly to the SDP transform rather than
-/// to a property-setter side-effect. Cross-check fields
-/// (`encoding-name`, video/ANC `clock-rate`, essence shape)
-/// don't pass through this builder — [`sdp::cross_check_essence`]
-/// reads them from `settings.transport_caps` directly.
-fn property_overrides_udp(settings: &CommonSettings) -> SdpOverrides<'_> {
-    fn opt(s: &str) -> Option<&str> {
-        if s.is_empty() { None } else { Some(s) }
-    }
-    fn opt_port(p: u16) -> Option<u16> {
-        if p == 0 { None } else { Some(p) }
-    }
-    let (source_ip, interface_ip, destination_ip, source_port) = match settings.side {
-        Side::Sender => (
-            opt(&settings.source_ip),
-            // Sender duplicates source_ip into the iface-ip slot
-            // — see the per-side dispatch note above.
-            opt(&settings.source_ip),
-            opt(&settings.destination_ip),
-            opt_port(settings.source_port),
-        ),
-        Side::Receiver => (
-            opt(&settings.source_ip),
-            opt(&settings.interface_ip),
-            opt(&settings.multicast_ip),
-            None,
-        ),
-    };
-    let tc = settings
-        .transport_caps
-        .as_ref()
-        .and_then(|c| c.structure(0));
-    // pt is i32 on `application/x-rtp` caps per GStreamer
-    // convention; cast to u8 for the [`SdpOverrides`] slot.
-    // 0..=255 keeps the cast lossless and lets the
-    // RFC-3551-range check fires centrally in
-    // `sdp::passthrough_with_overrides`.
-    let payload_type = tc
-        .and_then(|s| s.get::<i32>("payload").ok())
-        .and_then(|pt| u8::try_from(pt).ok());
-    let audio_clock_rate = tc
-        .and_then(|s| s.get::<i32>("clock-rate").ok())
-        .and_then(|rate| u32::try_from(rate).ok());
-    let a_ptime = tc.and_then(|s| s.get::<&str>("a-ptime").ok());
-    let a_maxptime = tc.and_then(|s| s.get::<&str>("a-maxptime").ok());
-
-    SdpOverrides {
-        label: opt(&settings.label),
-        description: opt(&settings.description),
-        name: opt(&settings.name),
-        interface_ip,
-        destination_ip,
-        destination_port: opt_port(settings.destination_port),
-        source_ip,
-        source_port,
-        payload_type,
-        audio_clock_rate,
-        a_ptime,
-        a_maxptime,
-        caps_mode: settings.caps_mode,
-    }
-}
-
-/// Build a [`FlowDefOverrides`] from the element's property snapshot.
-/// Empty-string properties map to `None` (i.e. "user did not set
-/// this; leave the file's value alone"). `mxl_domain_id` is taken
-/// from the domain resolution result, not the raw property, so the
-/// `domain_def.json`-derived value also flows into the splice when
-/// the user didn't set the property directly.
-fn property_overrides_mxl<'a>(
-    settings: &'a CommonSettings,
-    resolved_mxl_domain_id: &'a str,
-) -> FlowDefOverrides<'a> {
-    fn opt(s: &str) -> Option<&str> {
-        if s.is_empty() { None } else { Some(s) }
-    }
-    FlowDefOverrides {
-        flow_id: opt(&settings.mxl_flow_id),
-        label: opt(&settings.label),
-        description: opt(&settings.description),
-        name: opt(&settings.name),
-        mxl_domain_id: opt(resolved_mxl_domain_id),
-        caps_mode: settings.caps_mode,
-    }
-}
-
-fn log_flow_origin(cat: &gst::DebugCategory, field: &str, origin: ValueOrigin) {
-    match origin {
-        ValueOrigin::Property => gst::debug!(cat, "{field} from property; no transport file constraint"),
-        ValueOrigin::File => gst::info!(cat, "{field} taken from transport file"),
-        ValueOrigin::Both => gst::debug!(cat, "{field} cross-checked against transport file"),
-        ValueOrigin::None => gst::debug!(cat, "{field} not supplied by either source"),
-    }
-}
-
-/// Decide whether the element can build a real `mxlsink` / `mxlsrc`
-/// chain or has to fall back to its fake chain. Both sides need a
-/// non-empty Domain path and a non-empty flow id; the receiver
-/// additionally needs a specific [`FlowFormat`] (because `mxlsrc`
-/// has separate `video-flow-id` / `audio-flow-id` / `data-flow-id`
-/// properties).
-fn decide_inner_config_mxl(
-    settings: &CommonSettings,
-    flow: &flow_def::FlowResolution,
-    transport_file: Option<&str>,
-) -> InnerConfig {
-    if settings.mxl_domain_path.is_empty() {
-        return InnerConfig::Fake {
-            reason: "`mxl-domain-path` unset".to_owned(),
-        };
-    }
-    if flow.id.is_empty() {
-        return InnerConfig::Fake {
-            reason: "`mxl-flow-id` unset (neither property nor transport file supplied it)".to_owned(),
-        };
-    }
-    if settings.side == Side::Receiver && flow.format == FlowFormat::Unspecified {
-        return InnerConfig::Fake {
-            reason:
-                "`caps` media-type unrecognised or unset on nmossrc \
-                 (neither caps nor transport file pinned a flow format)"
-                    .to_owned(),
-        };
-    }
-    InnerConfig::Real(TransportConfig::Mxl {
-        domain_path: settings.mxl_domain_path.clone(),
-        flow_id: flow.id.clone(),
-        format: flow.format,
-        transport_file: transport_file.map(str::to_owned),
-    })
-}
-
-/// UDP sibling of [`decide_inner_config_mxl`]. Returns
-/// [`InnerConfig::Fake`] when there's no SDP to parse (deferred
-/// mode awaiting IS-05 PATCH); otherwise parses the SDP via
-/// [`sdp::parse_sdp`] and packages the resulting [`UdpMedia`] into
-/// a [`TransportConfig::Udp`]. SDP parse errors propagate as
-/// `Err` so the caller can ack-fail with attribution rather than
-/// silently downgrading to the fake chain (a malformed SDP is a
-/// real misconfiguration, not a "wait for more state to arrive"
-/// case).
-fn decide_inner_config_udp(
-    element: &str,
-    settings: &CommonSettings,
-    variant: UdpVariant,
-    transport_file: Option<&str>,
-    cross_check_mode: sdp::EssenceCrossCheckMode,
-) -> Result<InnerConfig, anyhow::Error> {
-    let Some(text) = transport_file else {
-        let reason = match settings.side {
-            Side::Sender => {
-                "no SDP transport file; waiting for IS-05 PATCH to supply the destination address"
-                    .to_owned()
-            }
-            Side::Receiver => {
-                "no SDP transport file; waiting for IS-05 PATCH to supply the listen address"
-                    .to_owned()
-            }
-        };
-        return Ok(InnerConfig::Fake { reason });
-    };
-    let media = sdp::parse_sdp(text).with_context(|| {
-        format!(
-            "{element}: parsing SDP transport file for transport={:?}",
-            settings.transport
-        )
-    })?;
-    // Cross-check the parsed SDP against the user-supplied
-    // `caps` (essence shape) and `transport-caps` (RTP-layer
-    // hints). The check fires after property overrides have
-    // applied the override-class fields, so an audio
-    // clock-rate that the user asked us to write into the
-    // SDP is implicit-OK while a video clock-rate disagreement
-    // (where clock-rate is cross-check, not override) surfaces
-    // as `SdpError::TransportCapsMismatch`. Mirrors
-    // `decide_inner_config_mxl`'s `resolve_mxl_flow_meta`
-    // cross-check pass.
-    sdp::cross_check_essence(
-        &media,
-        settings.caps.as_ref(),
-        settings.transport_caps.as_ref(),
-        cross_check_mode,
-    )
-    .with_context(|| {
-            format!(
-                "{element}: cross-checking SDP against `caps` / `transport-caps` \
-                 for transport={:?}",
-                settings.transport
-            )
-        })?;
-    Ok(InnerConfig::Real(TransportConfig::Udp {
-        variant,
-        media,
-        transport_file: Some(text.to_owned()),
-    }))
-}
-
-/// Cross-check strictness for [`decide_inner_config_udp`]: wide
-/// receivers (activation SDP carries `a=x-nvnmos-caps:`) relax
-/// essence shape on activation only.
-fn udp_essence_cross_check_mode(
-    settings: &CommonSettings,
-    activation: bool,
-    transport_file: Option<&str>,
-) -> sdp::EssenceCrossCheckMode {
-    if activation
-        && settings.side == Side::Receiver
-        && transport_file.is_some_and(sdp::indicates_wide_receiver_caps)
-    {
-        sdp::EssenceCrossCheckMode::FormatFamilyOnly
-    } else {
-        sdp::EssenceCrossCheckMode::Full
-    }
-}
-
-/// Transport-specific setup-time work for [`Transport::Mxl`]:
-/// domain-id resolution, transport-file synthesis from `caps`,
-/// property-overrides splice, and flow_def cross-checking. Extracted
-/// from [`validate_and_open`] so the top-level function reads as a
-/// clean transport dispatch.
-///
-/// Returns the resolved [`InnerConfig`] plus the transport-file
-/// text the daemon will be handed at `OpenSession` time (the
-/// post-synthesis, post-splice version of the input). The
-/// caller applies [`apply_auto_activate_gate`] separately.
-fn resolve_inner_config_mxl(
-    cat: &gst::DebugCategory,
-    element: &str,
-    settings: &CommonSettings,
-    resolved_transport_file: Option<String>,
-) -> Result<(InnerConfig, Option<String>), anyhow::Error> {
-    let domain_resolution =
-        domain::resolve_mxl_domain_id(&settings.mxl_domain_id, &settings.mxl_domain_path)
-            .with_context(|| format!("{element}: resolving MXL Domain identity"))?;
-    if domain_resolution.id.is_empty() {
-        bail!(
-            "{element}: `mxl-domain-id` is required when transport=mxl \
-             (set the property directly or supply an `mxl-domain-path` whose `domain_def.json` provides the id)"
-        );
-    }
-    match domain_resolution.origin {
-        DomainIdOrigin::Property => gst::debug!(
-            cat,
-            "mxl-domain-id from property; no `domain_def.json` consulted",
-        ),
-        DomainIdOrigin::DomainDef => gst::info!(
-            cat,
-            "mxl-domain-id taken from `domain_def.json` at `{}`",
-            settings.mxl_domain_path,
-        ),
-        DomainIdOrigin::Both => gst::debug!(
-            cat,
-            "mxl-domain-id cross-checked against `domain_def.json` at `{}`",
-            settings.mxl_domain_path,
-        ),
-        DomainIdOrigin::None => unreachable!("empty id rejected above"),
-    }
-
-    let transport_file = synthesise_or_passthrough_mxl(
-        cat,
-        element,
-        settings,
-        &domain_resolution.id,
-        resolved_transport_file,
-    )?;
-
-    // Property-overrides-file: splice any user-set identity/cosmetic
-    // properties (name, flow_id, mxl-domain-id, label, description,
-    // receiver-caps-mode) into the transport file before the daemon
-    // sees it. `caps` and `transport-caps` remain cross-checked by
-    // `resolve_mxl_flow_meta` below — they describe the essence
-    // shape and a mismatch is a real error.
-    let transport_file = match transport_file {
-        Some(text) => Some(
-            flow_def::splice_overrides(&text, &property_overrides_mxl(settings, &domain_resolution.id))
-                .with_context(|| format!("{element}: splicing property overrides into transport file"))?,
-        ),
-        None => None,
-    };
-
-    let caps_format = caps_format(settings);
-    let flow = flow_def::resolve_mxl_flow_meta(
-        &settings.mxl_flow_id,
-        caps_format,
-        transport_file.as_deref(),
-    )
-    .with_context(|| format!("{element}: resolving MXL flow id / format"))?;
-    log_flow_origin(cat, "mxl-flow-id", flow.id_origin);
-    log_flow_origin(cat, "caps format", flow.format_origin);
-
-    let mut inner = decide_inner_config_mxl(settings, &flow, transport_file.as_deref());
-    // Deferred-mode case (sender only): no resource is going to be
-    // registered at NULL→READY because neither `transport-file*` nor
-    // `caps` was supplied. Keep the fake chain so we don't bring
-    // `mxlsink` up against an unregistered Flow (which would fail to
-    // preroll); the inner is swapped to `mxlsink` only after
-    // `register_deferred` registers the Sender at READY→PAUSED.
-    if transport_file.is_none()
-        && settings.side == Side::Sender
-        && matches!(inner, InnerConfig::Real(_))
-    {
-        inner = InnerConfig::Fake {
-            reason: "deferred — peer caps will drive registration at READY\u{2192}PAUSED"
-                .to_owned(),
-        };
-    }
-
-    Ok((inner, transport_file))
-}
-
-/// Transport-specific setup-time work for [`Transport::Udp`] /
-/// [`Transport::Udp2`]: synthesise (or pass through) an SDP via
-/// [`synthesise_or_passthrough_udp`], splice property overrides
-/// via [`sdp::passthrough_with_overrides`] for user-supplied transport
-/// files, parse the result via
-/// [`sdp::parse_sdp`] inside [`decide_inner_config_udp`], and
-/// package the resulting [`UdpMedia`] into a
-/// [`TransportConfig::Udp`].
-///
-/// Precedence mirrors the MXL path:
-///
-/// * Explicit `transport-file*` → passthrough (caps cross-check
-///   against the file's essence shape via
-///   [`sdp::cross_check_essence`] in [`decide_inner_config_udp`]).
-/// * `caps` only (no `transport-file*`) → synthesise an SDP from
-///   `caps` + `transport_caps` + IS-05 endpoint properties +
-///   [`sdp::defaults`] via [`sdp::from_caps`].
-/// * Neither → no SDP; [`decide_inner_config_udp`] returns
-///   [`InnerConfig::Fake`] so the element waits for an IS-05
-///   PATCH to supply everything.
-///
-/// Activation-time SDP supersedes both at PATCH time
-/// (see [`make_activation_plan`]).
-fn resolve_inner_config_udp(
-    cat: &gst::DebugCategory,
-    element: &str,
-    settings: &CommonSettings,
-    variant: UdpVariant,
-    resolved_transport_file: Option<String>,
-) -> Result<(InnerConfig, Option<String>), anyhow::Error> {
-    // Synthesise an SDP from caps when no transport-file*
-    // is supplied; pass through the transport-file* otherwise.
-    // Mirrors `resolve_inner_config_mxl`'s
-    // `synthesise_or_passthrough_mxl` call.
-    let had_user_transport_file = resolved_transport_file.is_some();
-    let resolved_transport_file =
-        synthesise_or_passthrough_udp(cat, element, settings, resolved_transport_file)?;
-
-    // Property-overrides passthrough: rewrite any user-set
-    // identity / cosmetic / network properties (label,
-    // description, name, IS-05 endpoints, caps_mode) into
-    // user-supplied transport files before the daemon sees them.
-    // Mirrors `resolve_inner_config_mxl`'s
-    // `flow_def::splice_overrides` call. Activation-time SDP stays
-    // authoritative (see `make_activation_plan`) — the passthrough
-    // runs at startup only. Synthesised SDPs already bake every
-    // property in via [`sdp::from_caps`]; skip the second pass.
-    let resolved_transport_file = match resolved_transport_file {
-        Some(text) if had_user_transport_file => Some(
-            sdp::passthrough_with_overrides(&text, &property_overrides_udp(settings))
-                .with_context(|| {
-                    format!("{element}: applying property overrides to transport-file SDP")
-                })?,
-        ),
-        other => other,
-    };
-    let inner = decide_inner_config_udp(
-        element,
-        settings,
-        variant,
-        resolved_transport_file.as_deref(),
-        sdp::EssenceCrossCheckMode::Full,
-    )?;
-    Ok((inner, resolved_transport_file))
-}
-
 /// Honour `auto-activate` at setup time. When `auto_activate` is
 /// `false` and `decide_inner_config_mxl` / `decide_inner_config_udp`
 /// would have produced a real transport chain, downgrade to
@@ -1706,93 +1031,6 @@ pub(crate) fn make_activation_plan(
 
     ActivationPlan { inner, ack }
 }
-
-/// MXL branch of [`make_activation_plan`]: re-resolve domain id +
-/// flow meta against the activation's transport file, then run
-/// [`decide_inner_config_mxl`]. Returns `Ok(InnerConfig)` on
-/// success; on failure returns a fully-formed [`ActivationPlan`]
-/// (fake inner + failure ack) so the caller can short-circuit with
-/// the right error attribution.
-///
-/// The `Err` variant is boxed because the rare failure path's
-/// `ActivationPlan` (~240 B) would otherwise dominate the
-/// `Result`'s stack footprint on the common `Ok` path.
-fn resolve_activation_inner_mxl(
-    cat: &gst::DebugCategory,
-    element: &str,
-    settings: &CommonSettings,
-    transport_file: &str,
-) -> Result<InnerConfig, Box<ActivationPlan>> {
-    let domain_resolution =
-        match domain::resolve_mxl_domain_id(&settings.mxl_domain_id, &settings.mxl_domain_path) {
-            Ok(r) => r,
-            Err(e) => {
-                return Err(Box::new(ActivationPlan {
-                    inner: InnerConfig::Fake {
-                        reason: "mxl-domain-id resolution failed".to_owned(),
-                    },
-                    ack: ActivationAck::Failure {
-                        reason: format!(
-                            "{element}: resolving MXL Domain identity for activation: {e:#}"
-                        ),
-                    },
-                }));
-            }
-        };
-    if domain_resolution.id.is_empty() {
-        return Err(Box::new(ActivationPlan {
-            inner: InnerConfig::Fake {
-                reason: "mxl-domain-id unresolved".to_owned(),
-            },
-            ack: ActivationAck::Failure {
-                reason: format!(
-                    "{element}: activation rejected — `mxl-domain-id` is not resolvable on this \
-                     host (neither the property nor `mxl-domain-path`/`domain_def.json` \
-                     supplied an id)",
-                ),
-            },
-        }));
-    }
-    match domain_resolution.origin {
-        DomainIdOrigin::Property | DomainIdOrigin::DomainDef | DomainIdOrigin::Both => gst::debug!(
-            cat,
-            "{element}: activation mxl-domain-id resolved (origin={:?})",
-            domain_resolution.origin,
-        ),
-        DomainIdOrigin::None => unreachable!("empty id handled above"),
-    }
-
-    // Activation: the daemon's transport file is authoritative. Pass
-    // an empty `property_id` so the file always wins silently (the
-    // element's `mxl-flow-id` property is just a NULL→READY default;
-    // an IS-05 PATCH legitimately replaces it). The `caps` format
-    // cross-check stays because a v210 video activation arriving at
-    // an `nmossrc` configured for audio is a real misconfiguration
-    // the element must ack-fail.
-    let flow = match flow_def::resolve_mxl_flow_meta(
-        "",
-        caps_format(settings),
-        Some(transport_file),
-    ) {
-        Ok(r) => r,
-        Err(e) => {
-            return Err(Box::new(ActivationPlan {
-                inner: InnerConfig::Fake {
-                    reason: "flow_def resolution failed".to_owned(),
-                },
-                ack: ActivationAck::Failure {
-                    reason: format!(
-                        "{element}: resolving MXL flow id / format from activation \
-                         transport file: {e:#}"
-                    ),
-                },
-            }));
-        }
-    };
-
-    Ok(decide_inner_config_mxl(settings, &flow, Some(transport_file)))
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
